NahaJuliaLib.jl

NahaJuliaJib is a library of small utilities that I wish Julia had, but, near as I can tell, doesn't yet.

Type Hierarchy

NahaJuliaLib.allsubtypes — Function

allsubtypes(t::Type)

Return a Vector of t and all of its subtypes.

NahaJuliaLib.showsubtypes — Function

showsubtypes(t::Type) Print a hierarchical list of t and all of its subtypes.

NahaJuliaLib.pedigree — Function

pedigree(t::Type)

Return a Vector of t and its supertypes.

Defining Object Properties

Rather than use a condirional tree to determine which peoperty is being queried by getproperty, we can method specialize on Val types. We should then be able to automate propertynames based on the defined methods.

A struct might have fields, which are exposed as properties. The Val methods will not shadow the method that implements that.

NahaJuliaLib.@njl_getprop — Macro

@njl_getprop MyStruct

Define the methods necessary so that ValgetpropertiesofMyStructwill findVal` specialized properties.

Here'sa trivial, but illustrative example:

using NahaJuliaLib

struct MyStruct
    a::Int
end

@njl_getprop MyStruct

function Base.getproperty(o::MyStruct, prop::Val{:b})
    o.a * 2
end

ms = MyStruct(3)
MyStruct(3)

ms.a

ms.b

propertynames(ms)

2-element Vector{Symbol}:
 :a
 :b